re-organize chem utility
