+++ tdc/chem_utils/__init__.py
@@ -1,6 +1,6 @@
 from .evaluator import validity, uniqueness, novelty, diversity, kl_divergence, fcd_distance
-from .molconvert import MolConvert 
-from .oracle import docking_meta, molecule_one_retro, ibm_rxn, \
+from .featurize.molconvert import MolConvert 
+from .oracle.oracle import docking_meta, molecule_one_retro, ibm_rxn, \
 					askcos, isomers_c7h8n2o2, isomers_c9h10n2o2pf2cl, \
 					valsartan_smarts, scaffold_hop, deco_hop, \
 					sitagliptin_mpo, zaleplon_mpo, amlodipine_mpo, \
@@ -11,4 +11,4 @@ from .oracle import docking_meta, molecule_one_retro, ibm_rxn, \
 					troglitazone_rediscovery, thiothixene_rediscovery, \
 					median_meta, isomer_meta, rediscovery_meta, similarity_meta, \
 					jnk3, gsk3b, SA, cyp3a4_veith, drd2, qed, penalized_logp
-from .molfilter import MolFilter
+from .oracle.filter import MolFilter
