+++ tdc/chem_utils/featurize/_smartsPatts.py
