+++ tdc/chem_utils/featurize/_smiles2pubchem.py
@@ -43,8 +43,9 @@ try:
 except:
 	raise ImportError("Please install networkx by 'pip install networkx'! ")	
 
-from ..utils import oracle_load, print_sys, install
-from .smartsPatts import smartsPatts
+from tdc.utils import oracle_load, print_sys, install
+# from ..utils import oracle_load, print_sys, install
+from ._smartsPatts import smartsPatts
 
 PubChemKeys = None
