+++ tdc/chem_utils/featurize/_xyz2mol.py
@@ -43,7 +43,8 @@ try:
 except:
 	raise ImportError("Please install networkx by 'pip install networkx'! ")	
 
-from ..utils import print_sys 
+from tdc.utils import oracle_load, print_sys, install
+# from ..utils import print_sys
